# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Departments API",
        "description": "Documentación de la API de Departamentos con Swagger en Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
        "/api/departments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Obtener lista de departamentos",
                "description": "Listar todos los departamentos",
                "operationId": "da8820891dd6bf08713e9d77fbe6389f",
                "responses": {
                    "200": {
                        "description": "Lista de departamentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Human Resources"
                                            },
                                            "level": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "employee_count": {
                                                "type": "integer",
                                                "example": 12
                                            },
                                            "parent_department_id": {
                                                "type": "integer",
                                                "example": 1,
                                                "nullable": true
                                            },
                                            "ambassador_first_name": {
                                                "type": "string",
                                                "example": "John",
                                                "nullable": true
                                            },
                                            "ambassador_last_name": {
                                                "type": "string",
                                                "example": "Smith",
                                                "nullable": true
                                            },
                                            "is_active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-19T12:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-19T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Departments"
                ],
                "summary": "Crear departamento",
                "description": "Crear un nuevo departamento",
                "operationId": "c0f48115550ccf6a76b43ff263bfc05c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "level",
                                    "employee_count"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Finance"
                                    },
                                    "level": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "employee_count": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "parent_department_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "ambassador_first_name": {
                                        "type": "string",
                                        "example": "Mary",
                                        "nullable": true
                                    },
                                    "ambassador_last_name": {
                                        "type": "string",
                                        "example": "Johnson",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Departamento creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Finance"
                                        },
                                        "level": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "employee_count": {
                                            "type": "integer",
                                            "example": 8
                                        },
                                        "parent_department_id": {
                                            "type": "integer",
                                            "example": 1,
                                            "nullable": true
                                        },
                                        "ambassador_first_name": {
                                            "type": "string",
                                            "example": "Mary",
                                            "nullable": true
                                        },
                                        "ambassador_last_name": {
                                            "type": "string",
                                            "example": "Johnson",
                                            "nullable": true
                                        },
                                        "is_active": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/departments/{id}/subdepartments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Obtener subdepartamentos de un departamento",
                "description": "Listar subdepartamentos de un departamento",
                "operationId": "1678cb46bfa2bb8c6f0fdd9642727021",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del departamento padre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de subdepartamentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Finance Team"
                                            },
                                            "level": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "employee_count": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "parent_department_id": {
                                                "type": "integer",
                                                "example": 1,
                                                "nullable": true
                                            },
                                            "ambassador_first_name": {
                                                "type": "string",
                                                "example": "Anna",
                                                "nullable": true
                                            },
                                            "ambassador_last_name": {
                                                "type": "string",
                                                "example": "Taylor",
                                                "nullable": true
                                            },
                                            "is_active": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}